# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DWLite Open API",
        "description": "\u7403\u978b\u4f9b\u5e94\u94fe\u8d44\u6e90\u64ae\u5408\u4e0e\u6e20\u9053\u5206\u9500\u5e73\u53f0 Open API\n\n## \u8ba4\u8bc1\u65b9\u5f0f\n\n\u6240\u6709 API \u8bf7\u6c42\u9700\u8981\u4ee5\u4e0b Header:\n- `X-Api-Key`: API Key ID\n- `X-Timestamp`: Unix \u65f6\u95f4\u6233\uff08\u79d2\uff09\n- `X-Nonce`: \u552f\u4e00 UUID\uff08\u9632\u91cd\u653e\uff09\n- `X-Signature`: HMAC-SHA256 \u7b7e\u540d\n\n## \u7b7e\u540d\u7b97\u6cd5\n\n```\ncanonical = METHOD + \"\\n\" + PATH + \"\\n\" + TIMESTAMP + \"\\n\" + NONCE + \"\\n\" + SHA256(BODY)\nsignature = \"sha256=\" + HMAC-SHA256(canonical, API_SECRET)\n```\n\n## \u54cd\u5e94\u683c\u5f0f\n\n\u6210\u529f\u54cd\u5e94:\n```json\n{\n    \"success\": true,\n    \"data\": { ... },\n    \"requestId\": \"uuid\"\n}\n```\n\n\u9519\u8bef\u54cd\u5e94:\n```json\n{\n    \"success\": false,\n    \"error\": {\n        \"code\": \"ERROR_CODE\",\n        \"message\": \"\u9519\u8bef\u63cf\u8ff0\"\n    },\n    \"requestId\": \"uuid\"\n}\n```\n",
        "contact": {
            "name": "DWLite API Support"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1/open",
            "description": "Open API"
        }
    ],
    "paths": {
...
        "/api/v1/open/warehouse/outbound/orders": {
            "get": {
                "tags": [
                    "Warehouse - Outbound"
                ],
                "summary": "List outbound orders.",
                "operationId": "get_open_api_warehouse_outbound_list",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "picking",
                                "picked",
                                "packing",
                                "packed",
                                "shipped",
                                "cancelled"
                            ],
                            "nullable": true
                        }
                    },
                    {
                        "name": "createdFrom",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "createdTo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "outboundNo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 50,
                            "nullable": true
                        }
                    },
                    {
                        "name": "fulfillmentNo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 50,
                            "nullable": true
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/open/warehouse/outbound/orders/{outboundNo}": {
            "get": {
                "tags": [
                    "Warehouse - Outbound"
                ],
                "summary": "Get outbound order details.",
                "operationId": "get_open_api_warehouse_outbound_details",
                "parameters": [
                    {
                        "name": "outboundNo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/open/warehouse/outbound/orders/{outboundNo}/start-picking": {
            "post": {
                "tags": [
                    "Warehouse - Outbound"
                ],
                "summary": "Start picking outbound order.",
                "operationId": "post_open_api_warehouse_outbound_start_picking",
                "parameters": [
                    {
                        "name": "outboundNo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StartPickingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/open/warehouse/outbound/orders/{outboundNo}/complete-picking": {
            "post": {
                "tags": [
                    "Warehouse - Outbound"
                ],
                "summary": "Complete picking of outbound order.",
                "operationId": "post_open_api_warehouse_outbound_complete_picking",
                "parameters": [
                    {
                        "name": "outboundNo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompletePickingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/open/warehouse/outbound/orders/{outboundNo}/complete-packing": {
            "post": {
                "tags": [
                    "Warehouse - Outbound"
                ],
                "summary": "Complete packing of outbound order.",
                "operationId": "post_open_api_warehouse_outbound_complete_packing",
                "parameters": [
                    {
                        "name": "outboundNo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompletePackingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/open/warehouse/outbound/orders/{outboundNo}/ship": {
            "post": {
                "tags": [
                    "Warehouse - Outbound"
                ],
                "summary": "Ship outbound order.",
                "operationId": "post_open_api_warehouse_outbound_ship",
                "parameters": [
                    {
                        "name": "outboundNo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ShipOutboundRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StartPickingRequest": {
                "required": [
                    "startedAt"
                ],
                "properties": {
                    "startedAt": {
                        "type": "string"
                    },
                    "pickerName": {
                        "type": "string",
                        "maxLength": 100,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PickedItemDto": {
                "required": [
                    "sku",
                    "pickedQuantity"
                ],
                "properties": {
                    "sku": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "pickedQuantity": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "shelvingLocation": {
                        "type": "string",
                        "maxLength": 200,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "CompletePickingRequest": {
                "required": [
                    "items",
                    "completedAt"
                ],
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PickedItemDto"
                        },
                        "minItems": 1
                    },
                    "completedAt": {
                        "type": "string"
                    },
                    "notes": {
                        "type": "string",
                        "maxLength": 500,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "CompletePackingRequest": {
                "required": [
                    "completedAt"
                ],
                "properties": {
                    "completedAt": {
                        "type": "string"
                    },
                    "packerName": {
                        "type": "string",
                        "maxLength": 100,
                        "nullable": true
                    },
                    "packageWeight": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "nullable": true
                    },
                    "packageDimensions": {
                        "type": "string",
                        "maxLength": 50,
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "maxLength": 500,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ShipOutboundRequest": {
                "required": [
                    "shippedAt",
                    "shippingCarrier",
                    "trackingNumber"
                ],
                "properties": {
                    "shippedAt": {
                        "type": "string"
                    },
                    "shippingCarrier": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "trackingNumber": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "trackingUrl": {
                        "type": "string",
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "maxLength": 500,
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API Key ID",
                "name": "X-Api-Key",
                "in": "header"
            },
            "TimestampAuth": {
                "type": "apiKey",
                "description": "Unix \u65f6\u95f4\u6233\uff08\u79d2\uff09",
                "name": "X-Timestamp",
                "in": "header"
            },
            "NonceAuth": {
                "type": "apiKey",
                "description": "\u552f\u4e00 UUID\uff08\u9632\u91cd\u653e\u653b\u51fb\uff09",
                "name": "X-Nonce",
                "in": "header"
            },
            "SignatureAuth": {
                "type": "apiKey",
                "description": "HMAC-SHA256 \u7b7e\u540d",
                "name": "X-Signature",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "ApiKeyAuth": [],
            "TimestampAuth": [],
            "NonceAuth": [],
            "SignatureAuth": []
        }
    ],
...
}